# Fix unit test failures and improve test tolerance handling